+++ Bundles/ProductPackagingUnit/src/Spryker/Zed/ProductPackagingUnit/Business/Model/Reservation/LeadProductReservationCalculatorInterface.php
@@ -15,7 +15,7 @@ interface LeadProductReservationCalculatorInterface
      * @param string $leadProductSku
      * @param \Generated\Shared\Transfer\StoreTransfer $storeTransfer
      *
-     * @return int
+     * @return float
      */
-    public function calculateReservedAmountForLeadProduct(string $leadProductSku, StoreTransfer $storeTransfer): int;
+    public function calculateReservedAmountForLeadProduct(string $leadProductSku, StoreTransfer $storeTransfer): float;
 }
